Big refactor - support multi-tenant oauth; cleanup client/db interfaces
