+++ src/__tests__/auth.test.ts
@@ -1,8 +1,8 @@
 import { describe, it, expect } from 'vitest';
 import { requireOAuthUser } from '../auth.js';
 import httpMocks from 'node-mocks-http';
-import { SqliteOAuthClientDb } from '../oauthClientDb.js';
-import { OAuthResourceServerClient } from '../oauthResourceServerClient.js';
+import { SqliteOAuthDb } from '../oAuthDb.js';
+import { OAuthClient } from '../oAuthClient.js';
 import { mockAuthorizationServer } from './testHelpers';
 import fetchMock from 'fetch-mock';
 
@@ -11,10 +11,10 @@ describe('requireOAuthAuthUser', () => {
     const { req, res } = httpMocks.createMocks();
     const f = fetchMock.createInstance();
     mockAuthorizationServer(f, 'https://paymcp.com');
-    const db = new SqliteOAuthClientDb(':memory:');
-    const client = new OAuthResourceServerClient('https://paymcp.com', db, 'https://paymcp.com/callback', f.fetchHandler);
+    const db = new SqliteOAuthDb(':memory:');
+    const client = new OAuthClient("bdj", db, 'https://paymcp.com/callback', false, f.fetchHandler);
 
-    const fn = requireOAuthUser(client);
+    const fn = requireOAuthUser('https://paymcp.com', client);
     const user = await fn(req, res);
     expect(user).toBeUndefined();
     expect(res.statusCode).toEqual(401);
@@ -29,9 +29,9 @@ describe('requireOAuthAuthUser', () => {
     const res = httpMocks.createResponse();
     const f = fetchMock.createInstance();
     mockAuthorizationServer(f, 'https://paymcp.com');
-    const client = new OAuthResourceServerClient('https://paymcp.com', new SqliteOAuthClientDb(':memory:'), 'https://paymcp.com/callback', f.fetchHandler);
+    const client = new OAuthClient("bdj", new SqliteOAuthDb(':memory:'), 'https://paymcp.com/callback', false, f.fetchHandler);
 
-    const fn = requireOAuthUser(client);
+    const fn = requireOAuthUser('https://paymcp.com', client);
     const user = await fn(req, res);
     expect(user).toBeUndefined();
     expect(res.statusCode).toEqual(401);
@@ -48,9 +48,9 @@ describe('requireOAuthAuthUser', () => {
     const res = httpMocks.createResponse();
     const f = fetchMock.createInstance();
     mockAuthorizationServer(f, 'https://paymcp.com');
-    const client = new OAuthResourceServerClient('https://paymcp.com', new SqliteOAuthClientDb(':memory:'), 'https://paymcp.com/callback', f.fetchHandler);
+    const client = new OAuthClient("bdj", new SqliteOAuthDb(':memory:'), 'https://paymcp.com/callback', false, f.fetchHandler);
 
-    const fn = requireOAuthUser(client);
+    const fn = requireOAuthUser('https://paymcp.com', client);
     const user = await fn(req, res);
     expect(user).toBeUndefined();
     expect(res.statusCode).toEqual(401);
@@ -63,9 +63,9 @@ describe('requireOAuthAuthUser', () => {
     const res = httpMocks.createResponse();
     const f = fetchMock.createInstance();
     mockAuthorizationServer(f, 'https://paymcp.com');
-    const client = new OAuthResourceServerClient('https://paymcp.com', new SqliteOAuthClientDb(':memory:'), 'https://paymcp.com/callback', f.fetchHandler);
+    const client = new OAuthClient("bdj", new SqliteOAuthDb(':memory:'), 'https://paymcp.com/callback', false, f.fetchHandler);
 
-    const fn = requireOAuthUser(client);
+    const fn = requireOAuthUser('https://paymcp.com', client);
     const user = await fn(req, res);
     expect(user).toBeUndefined();
     expect(res.statusCode).toEqual(401);
@@ -78,9 +78,9 @@ describe('requireOAuthAuthUser', () => {
     const f = fetchMock.createInstance();
     mockAuthorizationServer(f, 'https://paymcp.com')
       .modifyRoute('https://paymcp.com/introspect', {method: 'post', response: {body: {active: false}}});
-    const client = new OAuthResourceServerClient('https://paymcp.com', new SqliteOAuthClientDb(':memory:'), 'https://paymcp.com/callback', f.fetchHandler);
+    const client = new OAuthClient("bdj", new SqliteOAuthDb(':memory:'), 'https://paymcp.com/callback', false, f.fetchHandler);
 
-    const fn = requireOAuthUser(client);
+    const fn = requireOAuthUser('https://paymcp.com', client);
     const user = await fn(req, res);
     expect(user).toBeUndefined();
     expect(res.statusCode).toEqual(401);
@@ -92,9 +92,9 @@ describe('requireOAuthAuthUser', () => {
     const f = fetchMock.createInstance();
     mockAuthorizationServer(f, 'https://paymcp.com')
       .modifyRoute('https://paymcp.com/introspect', {method: 'post', response: {body: {active: true, sub: 'test-user'}}});
-    const client = new OAuthResourceServerClient('https://paymcp.com', new SqliteOAuthClientDb(':memory:'), 'https://paymcp.com/callback', f.fetchHandler);
+    const client = new OAuthClient("bdj", new SqliteOAuthDb(':memory:'), 'https://paymcp.com/callback', false, f.fetchHandler);
 
-    const fn = requireOAuthUser(client);
+    const fn = requireOAuthUser('https://paymcp.com', client);
     const user = await fn(req, res);
     expect(user).toBe('test-user');
     expect(res.statusCode).toEqual(200);
@@ -105,17 +105,17 @@ describe('requireOAuthAuthUser', () => {
     const res = httpMocks.createResponse();
     const f = fetchMock.createInstance();
     mockAuthorizationServer(f, 'https://paymcp.com');
-    const db = new SqliteOAuthClientDb(':memory:');
+    const db = new SqliteOAuthDb(':memory:');
     const creds = {
       clientId: 'testClientId',
       clientSecret: 'testClientSecret',
       redirectUri: 'https://paymcp.com/callback'
     };
     db.saveClientCredentials('https://paymcp.com', creds);
-    const client = new OAuthResourceServerClient('https://paymcp.com', db, 'https://paymcp.com/callback', f.fetchHandler);
+    const client = new OAuthClient("bdj", db, 'https://paymcp.com/callback', false, f.fetchHandler);
 
     const encodedCreds = Buffer.from(`${encodeURIComponent(creds.clientId)}:${encodeURIComponent(creds.clientSecret)}`).toString('base64');
-    const fn = requireOAuthUser(client);
+    const fn = requireOAuthUser('https://paymcp.com', client);
     const user = await fn(req, res);
     expect(res.statusCode).toEqual(200);
     expect(user).toBe('testUser');
@@ -129,10 +129,10 @@ describe('requireOAuthAuthUser', () => {
     const res = httpMocks.createResponse();
     const f = fetchMock.createInstance();
     mockAuthorizationServer(f, 'https://paymcp.com');
-    const db = new SqliteOAuthClientDb(':memory:');
-    const client = new OAuthResourceServerClient('https://paymcp.com', db, 'https://paymcp.com/callback', f.fetchHandler);
+    const db = new SqliteOAuthDb(':memory:');
+    const client = new OAuthClient("bdj", db, 'https://paymcp.com/callback', false, f.fetchHandler);
 
-    const fn = requireOAuthUser(client);
+    const fn = requireOAuthUser('https://paymcp.com', client);
     const user = await fn(req, res);
     expect(res.statusCode).toEqual(200);
     expect(user).toBe('testUser');
@@ -145,38 +145,39 @@ describe('requireOAuthAuthUser', () => {
     const res = httpMocks.createResponse();
     const f = fetchMock.createInstance();
     mockAuthorizationServer(f, 'https://paymcp.com')
-      .modifyRoute('https://paymcp.com/introspect', {method: 'post', response: {status: 403, body: {}}})
-      .postOnce('https://paymcp.com/introspect', 
-        {
-          active: true,
-          sub: 'test-user'
-        });
-    const db = new SqliteOAuthClientDb(':memory:');
+      .modifyRoute('https://paymcp.com/introspect', {method: 'post', response: {status: 401, body: {}}})
+      .postOnce('https://paymcp.com/introspect', { 
+        active: true,
+        sub: 'testUser'
+      });
+    const db = new SqliteOAuthDb(':memory:');
     const creds = {
       clientId: 'oldClientId',
       clientSecret: 'oldClientSecret',
       redirectUri: 'https://paymcp.com/mcp/callback'
     };
-    db.saveClientCredentials('https://paymcp.com/mcp/callback', creds);
-    const client = new OAuthResourceServerClient('https://paymcp.com', db, 'https://paymcp.com/mcp/callback', f.fetchHandler);
+    db.saveClientCredentials('https://paymcp.com', creds);
+    const client = new OAuthClient("bdj", db, 'https://example.com/mcp/callback', false, f.fetchHandler);
 
-    const encodedCreds = Buffer.from(`${encodeURIComponent(creds.clientId)}:${encodeURIComponent(creds.clientSecret)}`).toString('base64');
-    const fn = requireOAuthUser(client);
+    const fn = requireOAuthUser('https://paymcp.com', client);
     const user = await fn(req, res);
     expect(res.statusCode).toEqual(200);
-    expect(user).toBe('test-user');
+    expect(user).toBe('testUser');
 
     // The first introspect call should have failed
     const introspectCalls = f.callHistory.calls('https://paymcp.com/introspect');
     expect(introspectCalls[0]).toBeDefined();
-    expect((introspectCalls[0].args[1] as any).headers['authorization']).toEqual(`Basic ${encodedCreds}`);
+    const firstIntrospectHeader = (introspectCalls[0].args[1] as any).headers['authorization'];
+    const decodedCreds = Buffer.from(firstIntrospectHeader.split(' ')[1], 'base64').toString('utf-8').split(':');
+    expect(decodedCreds[0]).toEqual(creds.clientId);
+    expect(decodedCreds[1]).toEqual(creds.clientSecret);
 
     // Then we should have registered a new client
     const registerCall = f.callHistory.lastCall('https://paymcp.com/register');
     expect(registerCall).toBeDefined();
 
     // And we should have updated the saved credentials
-    const savedCreds = await db.getClientCredentials('https://paymcp.com/mcp/callback');
+    const savedCreds = await db.getClientCredentials('https://paymcp.com');
     expect(savedCreds).toBeDefined();
     expect(savedCreds?.clientId).not.toEqual(creds.clientId);
     expect(savedCreds?.clientSecret).not.toEqual(creds.clientSecret);
