+++ src/auth.ts
@@ -1,5 +1,5 @@
 import { Request, Response } from "express";
-import { OAuthResourceServerClient } from './oauthResourceServerClient.js';
+import { OAuthGlobalClient } from "./oAuthGlobalClient.js";
 
 function getOp(req: Request): string {
   const isMessageEndpoint = req.path.endsWith('/message');
@@ -22,7 +22,7 @@ function getOp(req: Request): string {
 // opPrices is experimental: The names of tools that will be charged for if PayMcp is used. 
 // If not provided, all tools will be charged at the amount specified in the authorizationServerUrl's amount field
 // If any are provided, all unlisted tools will be charged at 0
-export function requireOAuthUser(oauthClient: OAuthResourceServerClient, opPrices?: {[key:string]: number}): (req: Request, res: Response) => Promise<string | undefined> {
+export function requireOAuthUser(tokenIntrospectionServerUrl: string, oauthClient: OAuthGlobalClient, opPrices?: {[key:string]: number}): (req: Request, res: Response) => Promise<string | undefined> {
   return async (req: Request, res: Response): Promise<string | undefined> => {
     const protocol = process.env.NODE_ENV === 'development' ? 'http' : 'https';
     const protectedResourceMetadataUrl = `${protocol}://${req.host}/.well-known/oauth-protected-resource${req.path}`;
@@ -56,7 +56,7 @@ export function requireOAuthUser(oauthClient: OAuthResourceServerClient, opPrice
         additionalParameters = { charge: opPrices[op] || 0 };
       }
       console.log('[auth] Introspecting token for op:', op, 'with additional parameters:', additionalParameters);
-      const introspectionResult = await oauthClient.introspectToken(token, additionalParameters);
+      const introspectionResult = await oauthClient.introspectToken(tokenIntrospectionServerUrl, token, additionalParameters);
       
       if (!introspectionResult.active) {
         console.log('[auth] Token is not active');
