+++ src/index.ts
@@ -1,7 +1,7 @@
 export * from './types.js';
 export * from './payMcpClient.js';
 export * from './solanaPaymentMaker.js';
-export * from './oauthClient.js';
-export * from './oauthResourceServerClient.js';
-export * from './oauthClientDb.js';
+export * from './oAuthClient.js';
+export * from './oAuthDb.js';
+export * from './oAuthGlobalClient.js';
 export * from './auth.js';
